requirements.txt
attrs==24.2.0
aws-cdk-lib==2.166.0
aws-cdk.asset-awscli-v1==2.2.210
aws-cdk.asset-kubectl-v20==2.1.3
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==38.0.1
blinker==1.8.2
cattrs==24.1.2
certifi==2024.8.30
click==8.1.7
constructs==10.4.2
Flask==3.0.3
gunicorn==23.0.0
h11==0.14.0
idna==3.10
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jsii==1.104.0
MarkupSafe==3.0.2
outcome==1.3.0.post0
packaging==24.1
publication==0.0.3
PySocks==1.7.1
python-dateutil==2.9.0.post0
selenium==4.25.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
trio==0.27.0
trio-websocket==0.11.1
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==2.2.3
websocket-client==1.8.0
Werkzeug==3.0.6
wsproto==1.2.0
